feat:unzip many file types |fix:Output when mode=1 Enable decompressing of various file types
The code for unpacking various compressed file types was there, but in
comments for some later release. Now this block of code is uncommented.

Output to correct dir when mode=1

Fixed the folder where the files are downloaded for the case when the
3rd argument (mode) equals 1. This is the expected behavior when user
gives as argument the desired destination folder.

Fix default output directory

The $destination_dir (2nd argument) defaulted to current directory which
was falsely designated as "./". This dosn't work in all environments (e.g.
inside WSL or VMs). A more robust solution is to use "$PWD", which works
like a charm in all tested cases.