feat: unzip files when mode=1 |fix: properly get files when mode=1 Properly download files when mode=1
Fixed the way individual files are downloaded by fixing the path to
cookies file. Since we change dir right before executing curl, we then
must enter the full path to the previous dir, where cookies file usually
resides. If cookies file is not provided, then the files downloaded are junk
(i.e. empty files - 0KB) or non-existent.

Unzip files when mode=1

After all links are correctly accessed and hence all files are downloaded, the
files that are in compressed form should be unpacked (and the initial files to
 be deleted).